meet cluster client.
