+++ aeron-cluster-poc-examples/src/main/java/io/scalecube/acpoc/Utils.java
@@ -1,4 +1,4 @@
-package poc;
+package io.scalecube.acpoc;
 
 import java.io.File;
 import java.util.UUID;
