+++ aeron-cluster-poc-examples/src/main/java/io/scalecube/acpoc/service/StubService.java
@@ -1,4 +1,4 @@
-package poc.service;
+package io.scalecube.acpoc.service;
 
 import io.aeron.Image;
 import io.aeron.Publication;
